feat: replaced connection between microservices with ActiveMQ
