+++ gym-service/src/main/resources/application-dev.yml
@@ -11,3 +11,6 @@ spring:
   jpa:
     hibernate:
       ddl-auto: update
+
+  activemq:
+    broker-url: "tcp://aws-ec2-url:7034"
