+++ training-tracker-service/src/main/java/org/tolking/config/ModelMapperConfig.java
@@ -1,0 +1,20 @@
+package org.tolking.config;
+
+import org.modelmapper.Condition;
+import org.modelmapper.Conditions;
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ModelMapperConfig {
+    @Bean
+    public ModelMapper modelMapper() {
+        ModelMapper modelMapper = new ModelMapper();
+        Condition<?, ?> skipNulls = Conditions.isNotNull();
+
+        modelMapper.getConfiguration().setPropertyCondition(skipNulls);
+
+        return modelMapper;
+    }
+}
