+++ training-tracker-service/src/main/java/org/tolking/enums/ActionType.java
@@ -1,0 +1,6 @@
+package org.tolking.enums;
+
+public enum ActionType {
+    ADD,
+    DELETE
+}
