+++ training-tracker-service/src/main/java/org/tolking/exception/ApiError.java
@@ -1,0 +1,11 @@
+package org.tolking.exception;
+
+import java.time.LocalDateTime;
+
+public record ApiError(
+        String path,
+        String message,
+        String status,
+        LocalDateTime localDateTime
+) {
+}
