+++ training-tracker-service/src/main/java/org/tolking/exception/ApiErrorWithoutMessage.java
@@ -1,0 +1,10 @@
+package org.tolking.exception;
+
+import java.time.LocalDateTime;
+
+public record ApiErrorWithoutMessage(
+        String path,
+        String status,
+        LocalDateTime localDateTime
+) {
+}
