+++ training-tracker-service/src/main/java/org/tolking/exception/TimeoutException.java
@@ -1,0 +1,7 @@
+package org.tolking.exception;
+
+public class TimeoutException extends RuntimeException {
+    public TimeoutException() {
+        super("Request timout. Please try again...");
+    }
+}
