+++ training-tracker-service/src/main/java/org/tolking/external_dto/TrainingEventDTO.java
