--- training-tracker-service/src/main/java/org/tolking/training_event_service/enums/ActionType.java
@@ -1,6 +1,0 @@
-package org.tolking.training_event_service.enums;
-
-public enum ActionType {
-    ADD,
-    DELETE
-}
