--- training-tracker-service/src/main/java/org/tolking/training_event_service/exception/ApiError.java
@@ -1,11 +1,0 @@
-package org.tolking.training_event_service.exception;
-
-import java.time.LocalDateTime;
-
-public record ApiError(
-        String path,
-        String message,
-        String status,
-        LocalDateTime localDateTime
-) {
-}
