--- training-tracker-service/src/main/java/org/tolking/training_event_service/exception/ApiErrorWithoutMessage.java
@@ -1,10 +1,0 @@
-package org.tolking.training_event_service.exception;
-
-import java.time.LocalDateTime;
-
-public record ApiErrorWithoutMessage(
-        String path,
-        String status,
-        LocalDateTime localDateTime
-) {
-}
