--- training-tracker-service/src/main/java/org/tolking/training_event_service/exception/TimeoutException.java
@@ -1,7 +1,0 @@
-package org.tolking.training_event_service.exception;
-
-public class TimeoutException extends RuntimeException {
-    public TimeoutException() {
-        super("Request timout. Please try again...");
-    }
-}
