+++ training-tracker-service/src/main/resources/application-local.yml
@@ -1,0 +1,9 @@
+spring:
+  config:
+    activate:
+      on-profile: "local"
+  activemq:
+    broker-url: tcp://localhost:61616
+    user: admin
+    password: admin
+    event-queue-name: training-event-queue
