+++ training-tracker-service/src/main/resources/application.yml
@@ -1,15 +1,19 @@
 spring:
   application:
     name: training-tracker-service
+
   datasource:
     url: jdbc:h2:mem:gym-club
     username: test
     password: test
+
   jooq:
     sql-dialect: org.hibernate.dialect.H2Dialect
+
   jpa:
     hibernate:
       ddl-auto: update
+
   h2:
     console:
       enabled: true
